+++ platform/esp32/main/board_config.h
@@ -3,6 +3,7 @@
 
 #ifdef MCUJS_BOARD_SEEED_RETERMINAL_STICKY
 #define MCUJS_BOARD_NAME "seeed_reterminal_sticky"
+#define MCUJS_HAS_SD 1
 #define MCUJS_LED_PIN 255
 #define MCUJS_HAS_TINYUF2 0
 #define MCUJS_FLASH_SIZE (32u * 1024u * 1024u)
@@ -25,6 +26,9 @@
 #define MCUJS_UART_TX_PIN 43 /* XIAO D6 */
 #define MCUJS_UART_RX_PIN 44 /* XIAO D7 */
 #define MCUJS_UART_BAUD 115200
+#ifndef MCUJS_HAS_SD
+#define MCUJS_HAS_SD 0
+#endif
 #define MCUJS_HAS_DVI 0
 #define MCUJS_HAS_NEOPIXEL 0